+++ src/modules/shared/components/Menu/Menu.tsx
@@ -53,7 +53,9 @@ const _Menu = ({ tabs, className }: MenuProps) => {
   const { pass } = useFlags();
 
   const userRoles = profile?.data.roles || [];
-  const isAdmin = userRoles.find((e) => e === 'admin' || e === 'superAdmin');
+  const isAdmin = Boolean(
+    userRoles.find((e) => e === 'admin' || e === 'superAdmin')
+  );
 
   useEffect(() => {
     const tabsDefault: TabsConfig[] = [
@@ -76,13 +78,6 @@ const _Menu = ({ tabs, className }: MenuProps) => {
         isVisible: true,
       },
       {
-        title: translate('components>menu>clients'),
-        icon: <DashIcon width={17} height={17} />,
-        link: PixwayAppRoutes.TOKENS_CLIENTS,
-        isVisible: !isProduction,
-        sub: true,
-      },
-      {
         title: translate('components>menu>wallet'),
         icon: <CardIcon width={17} height={17} />,
         link: PixwayAppRoutes.WALLET,
@@ -93,6 +88,13 @@ const _Menu = ({ tabs, className }: MenuProps) => {
         icon: <TicketIcon width={17} height={17} />,
         link: PixwayAppRoutes.TOKENPASS,
         isVisible: pass && isAdmin,
+      },
+      {
+        title: translate('components>menu>clients'),
+        icon: <DashIcon width={17} height={17} />,
+        link: PixwayAppRoutes.TOKENS_CLIENTS,
+        isVisible: !isProduction && isAdmin,
+        sub: true,
       },
       {
         title: translate('components>menu>integration'),