+++ src/modules/storefront/components/StorefrontPreview.tsx
@@ -2,9 +2,12 @@ import { useContext, useState } from 'react';
 import { useEffectOnce } from 'react-use';
 
 import { ThemeContext, ThemeProvider } from '../contexts';
-import { DefaultDataProps, TemplateData } from '../interfaces';
+import { TemplateData } from '../interfaces';
 import { Banner } from './Banner';
+import { Cookies } from './Cookies';
 import { Header } from './Header';
+import { Menu } from './Menu';
+import { Products } from './Products';
 
 export const StorefrontPreview = () => {
   return (
@@ -17,20 +20,20 @@ export const StorefrontPreview = () => {
 const Storefront = () => {
   const context = useContext(ThemeContext);
   const [currentTheme, setCurrentTheme] = useState<TemplateData | null>(null);
-  useEffectOnce(() => {
-    window.addEventListener('message', listener);
 
-    return () => window.removeEventListener('message', listener);
-  });
+  const listener = ({ data }: MessageEvent<TemplateData | string>) => {
+    if (typeof data === 'string') {
+      return context?.setPageName(data);
+    }
 
-  const safeOrigin = 'http://localhost:3000/';
-  const listener = (event: MessageEvent<TemplateData | string>) => {
-    if (event.origin !== safeOrigin) return;
-
-    if (typeof event.data === 'string') return context?.setPageName(event.data);
-
-    setCurrentTheme(event.data);
+    setCurrentTheme(data);
   };
+
+  useEffectOnce(() => {
+    addEventListener('message', listener);
+
+    return () => removeEventListener('message', listener);
+  });
 
   const data = { ...context?.pageTheme, ...currentTheme };
   const themeContext = context?.defaultTheme;
@@ -44,19 +47,28 @@ const Storefront = () => {
         return (
           <Component
             key={item.type + i}
-            data={item.props}
-            defaultData={
-              themeContext[item.type] as keyof DefaultDataProps['defaultData']
-            }
+            data={{ ...themeContext[item.type], ...item } as any}
           />
         );
       })}
+
+      <Copyright />
     </>
   );
 };
 
 const componentMap = {
   header: Header,
-  categories: () => <></>,
+  menu: Menu,
   banner: Banner,
+  products: Products,
+  cookies: Cookies,
+};
+
+const Copyright = () => {
+  return (
+    <div className="pw-w-full pw-grid pw-place-items-center pw-h-[42px] pw-bg-white pw-font-roboto pw-font-medium">
+      Copyright {new Date().getFullYear()} - [web/lock]
+    </div>
+  );
 };